requirements.txt
streamlit
tensorflow-cpu
Pillow
numpy
plotly
requests
streamlit-lottie